update
Updated to statistical tracking to ask students which school they are from.
Added hover boxes on scatter graphs, residuals, dot plots, and bootstrap confidence intervals so when you hover over a point it gives you details on the point and highlights it in the data table.
Rebuild of residuals graph into javascript, including by default not showing weighted average line, this is now an option in the graph options area.